Moved the claims checks up to the team plan review component.  Enabled work item management on the PI objective panel based on permissions.
